# dormitory-backend/static/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Dormitory API",
    "description": "API for managing dormitory operations including authentication, admin, areas, registrations, reports, rooms, users, contracts, bills, notifications, services, and payments",
    "version": "1.0.0"
  },
  "paths": {
...
    "/api/rooms": {
      "get": {
        "tags": ["Room"],
        "summary": "Get all rooms",
        "parameters": [
          {"name": "page", "in": "query", "schema": {"type": "integer", "default": 1}, "description": "Page number"},
          {"name": "limit", "in": "query", "schema": {"type": "integer", "default": 10}, "description": "Items per page"},
          {"name": "min_capacity", "in": "query", "schema": {"type": "integer"}, "description": "Minimum capacity"},
          {"name": "max_capacity", "in": "query", "schema": {"type": "integer"}, "description": "Maximum capacity"},
          {"name": "min_price", "in": "query", "schema": {"type": "number"}, "description": "Minimum price"},
          {"name": "max_price", "in": "query", "schema": {"type": "number"}, "description": "Maximum price"},
          {"name": "available", "in": "query", "schema": {"type": "boolean"}, "description": "Filter available rooms"},
          {"name": "search", "in": "query", "schema": {"type": "string"}, "description": "Search by name or description"},
          {"name": "area_id", "in": "query", "schema": {"type": "integer"}, "description": "Filter by area ID"}
        ],
        "responses": {
          "200": {
            "description": "List of rooms",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "rooms": {"type": "array", "items": {"$ref": "#/components/schemas/Room"}},
                    "total": {"type": "integer"},
                    "pages": {"type": "integer"},
                    "current_page": {"type": "integer"}
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/rooms/{room_id}": {
      "get": {
        "tags": ["Room"],
        "summary": "Get room by ID",
        "parameters": [
          {"name": "room_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Room ID"}
        ],
        "responses": {
          "200": {
            "description": "Room details",
            "content": {
              "application/json": {
                "schema": {"$ref": "#/components/schemas/Room"}
              }
            }
          },
          "404": {"description": "Room not found"}
        }
      }
    },
    "/api/admin/rooms": {
      "post": {
    "tags": ["Room"],
    "summary": "Create a new room with optional images (Admin)",
    "security": [{"Bearer": []}],
    "requestBody": {
      "required": true,
      "content": {
        "multipart/form-data": {
          "schema": {
            "type": "object",
            "properties": {
              "name": {
                "type": "string",
                "description": "Name of the room",
                "example": "Room A101"
              },
              "capacity": {
                "type": "integer",
                "description": "Maximum number of people the room can hold",
                "example": 4
              },
              "price": {
                "type": "number",
                "format": "float",
                "description": "Price of the room",
                "example": 500000.00
              },
              "area_id": {
                "type": "integer",
                "description": "ID of the area the room belongs to",
                "example": 1
              },
              "description": {
                "type": "string",
                "description": "Description of the room",
                "example": "Modern room with air conditioning",
                "nullable": true
              },
              "images": {
                "type": "array",
                "items": {
                  "type": "string",
                  "format": "binary"
                },
                "description": "Array of image files (png, jpg, jpeg, gif; max 20 files, 10MB each)",
                "example": ["room_a101_1.jpg", "room_a101_2.jpg"]
              },
              "alt_text_0": {
                "type": "string",
                "description": "Description for the first image",
                "example": "Room A101 overview",
                "nullable": true
              },
              "is_primary_0": {
                "type": "boolean",
                "description": "Whether the first image is primary",
                "example": true,
                "nullable": true
              },
              "sort_order_0": {
                "type": "integer",
                "description": "Sort order for the first image",
                "example": 0,
                "nullable": true
              }
            },
            "required": ["name", "capacity", "price", "area_id"]
          }
        }
      }
    },
    "responses": {
      "201": {
        "description": "Room created successfully",
        "content": {
          "application/json": {
            "schema": {
              "$ref": "#/components/schemas/Room"
            },
            "example": {
              "room_id": 1,
              "name": "Room A101",
              "capacity": 4,
              "price": "500000.00",
              "current_person_number": 0,
              "description": "Modern room with air conditioning",
              "status": "AVAILABLE",
              "area_id": 1,
              "area_details": {
                "area_id": 1,
                "name": "Building A"
              }
            }
          }
        }
      },
      "400": {
        "description": "Missing required fields, invalid images, or duplicate room name",
        "content": {
          "application/json": {
            "example": {
              "message": "Tên phòng đã tồn tại"
            }
          }
        }
      },
      "404": {
        "description": "Area not found",
        "content": {
          "application/json": {
            "example": {
              "message": "Không tìm thấy khu vực"
            }
          }
        }
      },
      "403": {
        "description": "Admin access required",
        "content": {
          "application/json": {
            "example": {
              "message": "Admin access required"
            }
          }
        }
      },
      "500": {
        "description": "Internal server error",
        "content": {
          "application/json": {
            "example": {
              "message": "Lỗi server không xác định"
            }
          }
        }
      }
    }
      }
    },
...
    "/api/admin/rooms/{room_id}/reports": {
      "get": {
        "tags": ["Room"],
        "summary": "Get reports of a room (Admin)",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "room_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Room ID"}
        ],
        "responses": {
          "200": {
            "description": "List of reports",
            "content": {
              "application/json": {
                "schema": {"type": "array", "items": {"$ref": "#/components/schemas/Report"}}
              }
            }
          },
          "404": {"description": "Room not found"},
          "403": {"description": "Admin access required"}
        }
      }
    },
    "/api/admin/rooms/{room_id}/users": {
      "get": {
        "tags": ["Room"],
        "summary": "Get users in a room (Admin)",
        "security": [{"Bearer": []}],
        "parameters": [
          {"name": "room_id", "in": "path", "required": true, "schema": {"type": "integer"}, "description": "Room ID"}
        ],
        "responses": {
          "200": {
            "description": "List of users",
            "content": {
              "application/json": {
                "schema": {"type": "array", "items": {"$ref": "#/components/schemas/User"}}
              }
            }
          },
          "404": {"description": "Room not found"},
          "403": {"description": "Admin access required"}
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "Area": {
        "type": "object",
        "properties": {
          "area_id": {"type": "integer"},
          "name": {"type": "string"}
        }
      },
...
      "Report": {
        "type": "object",
        "properties": {
          "report_id": {"type": "integer"},
          "room_id": {"type": "integer"},
          "user_id": {"type": "integer"},
          "report_type_id": {"type": "integer"},
          "content": {"type": "string"},
          "status": {"type": "string", "enum": ["PENDING", "RESOLVED", "REJECTED"]}
        }
      },
...
      "Room": {
        "type": "object",
        "properties": {
          "room_id": {
            "type": "integer",
            "example": 63
          },
          "name": {
            "type": "string",
            "example": "Room A"
          },
          "capacity": {
            "type": "integer",
            "example": 10
          },
          "price": {
            "type": "string",
            "example": "500000.00"
          },
          "description": {
            "type": "string",
            "example": "A cozy meeting room",
            "nullable": true
          },
          "area_id": {
            "type": "integer",
            "example": 1
          },
          "status": {
            "type": "string",
            "enum": ["AVAILABLE", "OCCUPIED", "MAINTENANCE", "DISABLED"],
            "example": "AVAILABLE"
          },
          "current_person_number": {
            "type": "integer",
            "example": 5
          },
          "area_details": {
            "$ref": "#/components/schemas/Area"
          }
        },
        "required": ["room_id", "name", "capacity", "price", "current_person_number", "status", "area_id"]
      },
...
     "User": {
        "type": "object",
        "properties": {
          "user_id": {
            "type": "integer",
            "example": 1
          },
          "fullname": {
            "type": "string",
            "example": "Nguyen Van A"
          },
          "email": {
            "type": "string",
            "format": "email",
            "example": "user@example.com"
          },
          "phone": {
            "type": "string",
            "example": "0123456789",
            "nullable": true
          },
          "date_of_birth": {
            "type": "string",
            "format": "date",
            "example": "2000-01-01",
            "nullable": true
          },
          "CCCD": {
            "type": "string",
            "example": "123456789012",
            "nullable": true
          },
          "class_name": {
            "type": "string",
            "example": "CNTT01",
            "nullable": true
          },
          "avatar_url": {
            "type": "string",
            "example": "/Uploads/avatars/user_example_com_nguyen_van_a/1_1634567890.jpg",
            "nullable": true
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "example": "2023-10-01T12:00:00Z",
            "nullable": true
          }
        },
        "required": ["user_id", "fullname", "email", "created_at"]
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}